powerai_serverside/requirements.txt
tensorflow
pillow
numpy
